+++ CHANGELOG.md
@@ -29,6 +29,9 @@
 - ["Write" button in component menu allows to evaluate it separately from the
   rest of the workflow][11523].
 - [The documentation editor can now display tables][11564]
+- [Table Input Widget is now matched for Table.input method instead of
+  Table.new. Values must be string literals, and their content is parsed to the
+  suitable type][11612].
 
 [11151]: https://github.com/enso-org/enso/pull/11151
 [11271]: https://github.com/enso-org/enso/pull/11271
@@ -48,6 +51,7 @@
 [11547]: https://github.com/enso-org/enso/pull/11547
 [11523]: https://github.com/enso-org/enso/pull/11523
 [11564]: https://github.com/enso-org/enso/pull/11564
+[11612]: https://github.com/enso-org/enso/pull/11612
 
 #### Enso Standard Library